Refactor the model and matching algorithm
Currently, we can only effectively use the `SemanticMatch` and
`EquivalenceTable` classes with the server. This structure was
not very effective for testing. We refactor away from using the
model and instead add a new `algorithm` module.

Semantic matches are now stored in a `networkx.DiGraph`, a
directed graph. The algorithm is performed on the graph with
effective graph operations, making the whole process more
efficient as well.

Furthermore, we clean up the service to use it in a more pythonic
way, eliminating the need for the `service_model` module.

Note, that this is a major refactor and solves multiple problems:

Fixes #1
Fixes #5