Fix docker container definition
After the major refactor, the Dockerfile was
broken. This fixes it, and adapts it to use
Python version 3.11 at the same time (as defined
in the `pyproject.toml`).

Furthermore, this adds a `compose.yaml` for easier
deployment.

Sadly, however, this introduces #15, as currently
we cannot run the container directly via command,
as there is a broken path definition.